+++ Rubbish_Bookstore/Finance.hpp
@@ -3,5 +3,5 @@
 #ifndef Finance_hpp
 #define Finance_hpp
 void ShowFinance(int n); // 该函数用于显示财务信息。当n为0时，全部信息。
-void AddDeal(double value); // 该函数用于添加交易记录。带有符号。
+void AddDeal(long double value); // 该函数用于添加交易记录。带有符号。
 #endif
